refactor code and tests
