+++ src/components/Spreadsheet/logic/useTableEvents.ts
@@ -1,4 +1,5 @@
 import { useClipboard } from '@/components/Spreadsheet/logic/useClipboard'
+import { useDraggable } from '@/components/Spreadsheet/logic/useDraggable'
 import { useKeyPress } from '@/components/Spreadsheet/logic/useKeyPress'
 import { useMouse } from '@/components/Spreadsheet/logic/useMouse'
 import { useSelection } from '@/components/Spreadsheet/logic/useSelection'
@@ -9,6 +10,7 @@ export function useTableEvents() {
   useClipboard(selectedElements)
   useKeyPress(selectedElements, removeSelection, addSelectionArea)
   useMouse(selectedElements, addSelectionArea, removeSelection)
+  useDraggable(selectedElements, removeSelection)
 
   return {
     ...selectors,
